requirements.txt
# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Linting
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0

# Type checking
mypy>=1.5.1

# Project specific
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
